% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btps.R
\docType{data}
\name{btps_df}
\alias{btps_df}
\title{BTPS Correction factor Lookup Table}
\format{
A data frame that has 2 columns:
\itemize{
\item \strong{Gas_temp_c}: Gas temperature in celsius
\item \strong{Factor_37}: factor to convert volume of gas to 37 celsius saturated
}
}
\usage{
btps_df
}
\description{
A data frame that has 2 columns:
}
\keyword{datasets}
