% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met-rate.R
\name{get_BSA}
\alias{get_BSA}
\title{Calculate Body Surface Area}
\usage{
get_BSA(wt_kg, ht_cm)
}
\arguments{
\item{wt_kg}{(Numeric) Weight in kilogram}

\item{ht_cm}{(Numeric) Height in centimetre}
}
\value{
Print report to console and return BSA invisibly
}
\description{
This function calculate BSA (Body Surface Area) in square metre using \href{http://www-users.med.cornell.edu/~spon/picu/calc/bsacalc.htm}{DuBois & DuBois formula} (DuBois D, DuBois EF)
}
\examples{

# BSA in square metre
get_BSA(wt_kg = 70, ht_cm = 180)
# Input can be vectorized
get_BSA(wt_kg = 70:72, ht_cm = 180:182)

}
