% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met-rate.R
\name{get_metabolic_rate}
\alias{get_metabolic_rate}
\title{Calculate Metabolic Rate}
\usage{
get_metabolic_rate(
  x,
  y,
  paper_speed = 25,
  baro,
  temp_c,
  wt_kg,
  ht_cm,
  cal_eqi_oxygen = 4.825
)
}
\arguments{
\item{x}{Displacement of tracing in x-direction (in millimeter).}

\item{y}{Displacement of tracing in y-direction (in millimeter).}

\item{paper_speed}{Paper speed of the kymograph (in millimeter/minute), 25 ml/min by default.}

\item{baro}{(If \code{condition} = "STPD") Barometric pressure at the recording site.}

\item{temp_c}{(If \code{condition} = "STPD") Temperature in celsius at the recording site.}

\item{wt_kg}{(Numeric) Weight of the subject in kilogram}

\item{ht_cm}{(Numeric) Height of the subject in centimetre}

\item{cal_eqi_oxygen}{(Numeric) Caloric equivalent of Oxygen (default is 4.825, at RQ = 0.82)}
}
\value{
Report printed to console, return metabolic rate in Cal/m2/hr, invisibly
}
\description{
This function compute metabolic rate (Cal/m2/hr) from \href{https://www.somatco.com/Recording-Spirometer-50-1833-50-1817.pdf}{Harvard Spirometer} tracing,
subject's weight and height, and environmental condition i.e., barometric pressure and temperature.
The output printed to console will show each step of calculation which is suitable for educational purposes.
}
\examples{
get_metabolic_rate(x = 20, y = 10,
                   baro = 760, temp_c = 25,
                   wt_kg = 80, ht_cm = 180)
}
