% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxycons.R
\name{get_oxycons}
\alias{get_oxycons}
\title{Calculate Oxygen Consumption from Harvard Spirometer}
\usage{
get_oxycons(
  x,
  y,
  paper_speed = 25,
  unit = c("L/hr", "ml/min"),
  condition = c("ATPS", "STPD"),
  baro = NA,
  temp_c = NA
)
}
\arguments{
\item{x}{Displacement of tracing in x-direction (in millimeter).}

\item{y}{Displacement of tracing in y-direction (in millimeter).}

\item{paper_speed}{Paper speed of the kymograph (in millimeter/minute).}

\item{unit}{(Character) Unit of the oxygen consumption to return, e.g. L/hr (default).}

\item{condition}{(Character) Environmental condition to calculate oxygen consumption that affect gas volume, must be one of:
\itemize{
\item \strong{ATPS:} Volume of gas at ambient (A) temperature (T) (room temperature) and barometric pressure (P) saturated (S) with water vapor.
\item \strong{STPD:} Volume of gas at the standard (S) temperature (T) of 0°C and a barometric pressure (P) of 760 mmHg, and in a dry state (D).
}}

\item{baro}{(If \code{condition} = "STPD") Barometric pressure at the recording site.}

\item{temp_c}{(If \code{condition} = "STPD") Temperature in celsius at the recording site.}
}
\value{
Print information to console and return oxygen consumption as numeric vector, invisibly.
}
\description{
This function compute oxygen consumption from \href{https://www.somatco.com/Recording-Spirometer-50-1833-50-1817.pdf}{Harvard Spirometer} tracing.
Simply define 2 points along oxygen line, and provide displacement in x-and y-direction.
Oxygen consumption can be reported in units: L/hr or ml/min, and at condition: ATPS or STPD.
}
\examples{
# Oxygen Consumption at ATPS in L/hr
get_oxycons(x = 100, y = 50)
# Oxygen Consumption at ATPS in ml/min
get_oxycons(x = 100, y = 50, unit = "ml/min")
# Oxygen Consumption at STPD, in L/hr, must provide `baro` and `temp_c`
get_oxycons(x = 100, y = 50, condition = "STPD", baro = 760, temp_c = 23)
}
