% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btps.R
\name{lung_vol_atps_btps}
\alias{lung_vol_atps_btps}
\title{Convert Lung Volume from ATPS to BTPS}
\usage{
lung_vol_atps_btps(
  temp = NA,
  FEV1 = NA,
  FVC = NA,
  PEF = NA,
  TV = NA,
  IC = NA,
  EC = NA,
  VC = NA
)
}
\arguments{
\item{temp}{(Numeric) Room Temperature in celsius when gas was collected.}

\item{FEV1}{(numeric) Forced Expiratory Volume in 1 second (L).}

\item{FVC}{(numeric) Forced Vital Capacity (L)}

\item{PEF}{(numeric) Peak Expiratory Flow (L/min)}

\item{TV}{(numeric) Tidal Volume (L)}

\item{IC}{(numeric) Inspiratory Capacity (L)}

\item{EC}{(numeric) Expiratory Capacity (L)}

\item{VC}{(numeric) Vital Capacity (L)}
}
\value{
A Tibble
}
\description{
This function converts several lung volume parameters at ATPS (Ambient Temperature and Pressure Saturated) to
lung volume at BTPS (Body Temperature, Pressure, water vapor Saturated).
}
\examples{
lung_vol_atps_btps(temp = 30, FEV1 = 5, FVC = 10, PEF = 4, TV = 9, IC = 10, EC = 12, VC = 10)
}
