% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{new_HarvardTracing}
\alias{new_HarvardTracing}
\title{New HarvardTracing Object}
\usage{
new_HarvardTracing(x, tracing_info = list())
}
\arguments{
\item{x}{A data frame with at least \code{x} and \code{y} column}

\item{tracing_info}{list of values to be assigned to \code{tracing_info} attribute}
}
\value{
a \code{HarvardTracing} data frame
}
\description{
New HarvardTracing Object
}
