% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslab-package.R
\docType{package}
\name{rslab-package}
\alias{rslab}
\alias{rslab-package}
\title{rslab: Functions for Respiratory System Physiology Lab}
\description{
This package contains functions related to respiratory system physiology such as BTPS calculator and Metabolic rate calculator.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lightbridge-KS/rslab}
  \item Report bugs at \url{https://github.com/Lightbridge-KS/rslab/issues}
}

}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
