% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracing.R
\name{sim_Harvard_tracing}
\alias{sim_Harvard_tracing}
\title{Simulate Harvard Spirometer Tracing}
\usage{
sim_Harvard_tracing(
  f = "cos",
  t_start = 0,
  t_end = 1,
  paper_speed = 25,
  TV = 500,
  RR = 20,
  oxycons = 0,
  oxycons_unit = c("L/hr", "ml/min"),
  y_int_O2_line = 0,
  seq_x_by = 0.05,
  epsilon_sd = NULL,
  seed = 1
)
}
\arguments{
\item{f}{(Character or Symbol) A function to generate 1 respiratory cycle (i.e., inspiration and expiration). The default, currently, is a cosine function. Any function can be specified such that the amplitude is 1 (peak at 1, trough at -1) and the wavelength is \code{2*pi}.}

\item{t_start}{(Numeric) Time point (minute) when the experiment start}

\item{t_end}{(Numeric) Time point (minute) when the experiment end}

\item{paper_speed}{(Numeric) Paper speed (mm/minute) of the tracing}

\item{TV}{(Numeric) Tidal volume (mL) that subject breath}

\item{RR}{(Numeric) Respiratory Rate (/min) of subject}

\item{oxycons}{(Numeric) Oxygen consumption of subject}

\item{oxycons_unit}{(Character) Units of the \code{oxycons} must be one of: "L/hr" or "ml/min".}

\item{y_int_O2_line}{(Numeric) Y-intercept of an oxygen-line, i.e., the baseline that passed through the trough of each respiratory cycle waves.}

\item{seq_x_by}{(Numeric) Difference between consecutive \code{x} values to generate. (similar to \code{by} argument of \code{seq()})}

\item{epsilon_sd}{If provided as numeric, it is a standard deviation of an error term sampled from Gaussian distribution with \code{mean = 0}}

\item{seed}{(Numeric) If provided \code{epsilon_sd}, It is a seed to generated random error variation.}
}
\value{
A data.frame with columns:
\itemize{
\item \strong{x}: x-axis data of the respiratory wave tracing (in milimeter)
\item \strong{y}: y-axis data of the respiratory wave tracing (in milimeter)
\item \strong{y_O2_line}: y-axis data of the oxygen-line (in milimeter)
}
}
\description{
This function simulates volume-time tracing data produced by breathing of a hypothetical subject as recorded by the Harvard spirometer.
A number of parameters can be specified to adjust the tracing data. The output of this function is a data frame to be plotted by the user.
}
