% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpd.R
\docType{data}
\name{stpd_df}
\alias{stpd_df}
\title{STPD Correction factor Lookup Table}
\format{
A data frame with 738 rows and 3 variables:
\describe{
\item{baro}{Barometric pressure in mmHG}
\item{temp_c}{Temperature in degree celsius}
\item{stpd_factor}{STPD correction factor}

}
}
\source{
William D. McArdle, Frank I. Katch, and Victor L. Katch (1996). Appendixes. In:Donna Balado Exercise Physiology:Energy, Nutrition, and Human Performance. MD, USA:Willams & Wilkins. P768
}
\usage{
stpd_df
}
\description{
A data frame contain STPD correction factor (in long format)
}
\keyword{datasets}
