% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpd.R
\docType{data}
\name{stpd_df_wide}
\alias{stpd_df_wide}
\title{STPD Correction factor Lookup Table (Wide format)}
\format{
A data frame with 41 rows and 19 variables:
\describe{
\item{baro}{Barometric pressure in mmHG}
\item{temp_c_15}{STPD correction factors at temperature of 15 degree celsius}
\item{temp_c_16}{STPD correction factors at temperature of 16 degree celsius}
\item{temp_c_17}{STPD correction factors at temperature of 17 degree celsius}
\item{temp_c_18}{STPD correction factors at temperature of 18 degree celsius}
\item{temp_c_19}{STPD correction factors at temperature of 19 degree celsius}
\item{temp_c_20}{STPD correction factors at temperature of 20 degree celsius}
\item{temp_c_21}{STPD correction factors at temperature of 21 degree celsius}
\item{temp_c_22}{STPD correction factors at temperature of 22 degree celsius}
\item{temp_c_23}{STPD correction factors at temperature of 23 degree celsius}
\item{temp_c_24}{STPD correction factors at temperature of 24 degree celsius}
\item{temp_c_25}{STPD correction factors at temperature of 25 degree celsius}
\item{temp_c_26}{STPD correction factors at temperature of 26 degree celsius}
\item{temp_c_27}{STPD correction factors at temperature of 27 degree celsius}
\item{temp_c_28}{STPD correction factors at temperature of 28 degree celsius}
\item{temp_c_29}{STPD correction factors at temperature of 29 degree celsius}
\item{temp_c_30}{STPD correction factors at temperature of 30 degree celsius}
\item{temp_c_31}{STPD correction factors at temperature of 31 degree celsius}
\item{temp_c_32}{STPD correction factors at temperature of 32 degree celsius}
...
}
}
\source{
William D. McArdle, Frank I. Katch, and Victor L. Katch (1996). Appendixes. In:Donna Balado Exercise Physiology:Energy, Nutrition, and Human Performance. MD, USA:Willams & Wilkins. P768
}
\usage{
stpd_df_wide
}
\description{
A data frame contain STPD correction factor (in wide format)
}
\keyword{datasets}
