% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{tracing_info}
\alias{tracing_info}
\title{Get Tracing Info from Harvard Spirometer Tracing}
\usage{
tracing_info(x)
}
\arguments{
\item{x}{An object class \code{HarvardTracing}}
}
\value{
A list with each element contains key information of the tracing
}
\description{
This function retrieve \code{tracing_info} attributes from \code{HarvardTracing} data frame
produced by \code{\link[=sim_Harvard_tracing]{sim_Harvard_tracing()}}.
}
